Split evm_native into multiple modules
